Rename monitor related files
