+++ Project1.vcxproj
@@ -17,7 +17,7 @@
     <ClCompile Include="camera.cpp" />
     <ClCompile Include="levels\monitor\monitor_camera.cpp" />
     <ClCompile Include="levels\monitor\monitor_level.cpp" />
-    <ClCompile Include="levels\monitor\skybox.cpp" />
+    <ClCompile Include="levels\monitor\monitor_skybox.cpp" />
     <ClCompile Include="level.cpp" />
     <ClCompile Include="levels\motherboard\motherboard_level.cpp" />
     <ClCompile Include="main.cpp" />
@@ -37,7 +37,7 @@
     <ClInclude Include="include\utility.h" />
     <ClInclude Include="levels\monitor\monitor_camera.h" />
     <ClInclude Include="levels\monitor\monitor_level.h" />
-    <ClInclude Include="levels\monitor\skybox.h" />
+    <ClInclude Include="levels\monitor\monitor_skybox.h" />
     <ClInclude Include="levels\motherboard\motherboard_level.h" />
   </ItemGroup>
   <ItemGroup>
