+++ levels/monitor/monitor_level.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "../../include/level.h"
-#include "skybox.h"
+#include "monitor_skybox.h"
 #include "monitor_camera.h"
 
 class Monitor: public Level {
